requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-redis==5.4.0
celery==5.3.4